Add navigation and move import to a fragment
